+++ js/dataUsers.js
@@ -1,0 +1,32 @@
+const NAMES = [
+  'Алексей',
+  'Артем',
+  'Юлия',
+  'Дмитрий',
+  'Станислав',
+  'Виталий',
+  'Юрий',
+  'Наталья',
+  'Галина',
+  'Светлана',
+];
+
+const COMMENTS = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра.',
+  'В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают.',
+  'Как можно было поймать такой неудачный момент?!',
+];
+
+const PHOTO_DESCRIPTION = [
+  'Отличное фото!',
+  'Зацените эффекты',
+  'Был в отпуске',
+  'Живое фото получилось',
+  'Это мой обед',
+  'Как вам мой новый костюм',
+];
+export {NAMES, COMMENTS, PHOTO_DESCRIPTION};
