+++ js/random-array-element.js
@@ -1,0 +1,7 @@
+// Функция  возвращающая случайное значение
+let getRandomArrayElement = (array) => {
+  let randomNumber = Math.floor(Math.random() * array.length);
+  return array[randomNumber];
+}
+
+export{getRandomArrayElement};
